+++ src/main/kotlin/io/github/krozov/detekt/koin/KoinRuleSetProvider.kt
@@ -3,6 +3,7 @@ package io.github.krozov.detekt.koin
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.RuleSet
 import io.gitlab.arturbosch.detekt.api.RuleSetProvider
+import io.github.krozov.detekt.koin.moduledsl.ConstructorDslAmbiguousParameters
 import io.github.krozov.detekt.koin.moduledsl.DeprecatedKoinApi
 import io.github.krozov.detekt.koin.moduledsl.DuplicateBindingWithoutQualifier
 import io.github.krozov.detekt.koin.moduledsl.EmptyModule
@@ -10,6 +11,7 @@ import io.github.krozov.detekt.koin.moduledsl.EnumQualifierCollision
 import io.github.krozov.detekt.koin.moduledsl.GenericDefinitionWithoutQualifier
 import io.github.krozov.detekt.koin.moduledsl.MissingScopedDependencyQualifier
 import io.github.krozov.detekt.koin.moduledsl.ModuleIncludesOrganization
+import io.github.krozov.detekt.koin.moduledsl.ParameterTypeMatchesReturnType
 import io.github.krozov.detekt.koin.moduledsl.SingleForNonSharedDependency
 import io.github.krozov.detekt.koin.moduledsl.UnassignedQualifierInWithOptions
 import io.github.krozov.detekt.koin.scope.FactoryInScopeBlock
@@ -66,6 +68,8 @@ public class KoinRuleSetProvider : RuleSetProvider {
                 DuplicateBindingWithoutQualifier(config),
                 GenericDefinitionWithoutQualifier(config),
                 EnumQualifierCollision(config),
+                ConstructorDslAmbiguousParameters(config),
+                ParameterTypeMatchesReturnType(config),
                 // Scope Management Rules
                 MissingScopeClose(config),
                 ScopedDependencyOutsideScopeBlock(config),